specs/phase-2/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlmodel==0.0.16
pydantic==2.5.0
pydantic-settings==2.1.0
PyJWT==2.8.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
asyncpg==0.29.0
passlib==1.7.4
bcrypt==4.1.0
cryptography>=41.0.0
python-jose[cryptography]>=3.3.0